% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/burning.R
\name{burning}
\alias{burning}
\title{Apply burning process to Markov Chain Monte Carlo simualtions by disgarding the begining period of sequences.}
\usage{
burning(sris, percnt)
}
\arguments{
\item{percnt}{}
}
\value{

}
\description{
Apply burning process to Markov Chain Monte Carlo simualtions by disgarding the begining period of sequences.
}
