% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mcmc_posterior.R
\name{mcmc_posterior}
\alias{mcmc_posterior}
\title{Markov Chain Monte Carlo (MCMC) simulations.}
\usage{
mcmc_posterior(cur_gid, saveResult = T)
}
\arguments{
\item{saveResult}{}
}
\value{

}
\description{
Markov Chain Monte Carlo (MCMC) simulations.
}
