% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/multinoulli_ratio.R
\name{multinoulli_ratio}
\alias{multinoulli_ratio}
\title{Calculate Multinoulli probabilities.}
\usage{
multinoulli_ratio(miu, alpha, gam, gmodel = "add0")
}
\arguments{
\item{gmodel}{}
}
\value{

}
\description{
Calculate Multinoulli probabilities.
}
