% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/thining.R
\name{thining}
\alias{thining}
\title{Apply thinning to Markov Chain Monte Carlo simulations.}
\usage{
thining(sris, thinby = 1)
}
\arguments{
\item{thinby}{}
}
\value{

}
\description{
Apply thinning to Markov Chain Monte Carlo simulations.
}
