% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Update_b.R
\name{update_b}
\alias{update_b}
\title{Update probability parameter b.}
\usage{
update_b()
}
\value{

}
\description{
The next state of parameter b is updated, based on the most current state of other parameters.
}
