% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Update_G.R
\name{update_gstatus}
\alias{update_gstatus}
\title{Update gene segment status.}
\usage{
update_gstatus()
}
\value{

}
\description{
The next state of gene segment status is updated, based on the most current state of other parameters.
}
