% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Update_H.R
\name{update_H}
\alias{update_H}
\title{Update SNP status parameter H.}
\usage{
update_H()
}
\value{

}
\description{
For J SNPs on the gene segment, iteratively update SNP status (H_1, H_2, ...., H_J).
Each SNP status is updated based on the most current state of all other parameters.
}
