% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/MCMC_iteration.R
\name{update_mcmc}
\alias{update_mcmc}
\title{MCMC iteration.}
\usage{
update_mcmc(itt, initl = NA)
}
\arguments{
\item{itt}{iteration numbers}

\item{initl}{initialization number}
}
\value{
esti.list:  a list of estimation objects
}
\description{
MCMC iteration.
}
\details{
Global objects:
ssample: a matrix \link{snp.nums * 9}
snp.nums: number of SNPs for analysis
iter: current status (Gibbs sampler)
}
