% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Update_RR.R
\name{update_risk}
\alias{update_risk}
\title{Update relative risk parameter.}
\usage{
update_risk()
}
\value{

}
\description{
For J SNPs on the segment, iteratively update relative risk parameter (R_1, ..., R_J). The next state is updated based on the most current state of other parameters.
}
